Make CPP work in damldocs.
