+++ compiler/damlc/daml-opts/daml-opts/DA/Daml/Options.hs
@@ -29,6 +29,7 @@ import HscMain
 import Panic (throwGhcExceptionIO)
 import System.Directory
 import System.FilePath
+import System.Info.Extra (isWindows)
 
 import DA.Daml.Options.Types
 import DA.Daml.Preprocessor
@@ -253,6 +254,8 @@ adjustDynFlags options@Options{..} dflags
                 -- pragmas, line numbers may be wrong up when using CPP.
                 -- Ideally we fix the issue with the daml parser and
                 -- then remove this flag.
+            , sTmpDir = if isWindows then "C:\\Windows\\Temp" else "/tmp"
+                -- awful hack ... sometimes this is required by CPP?
             }
 
     -- We need to add platform info in order to run CPP. To prevent
